added BreakpointModulator to RAPT

+++ Libraries/RAPT/Source/Executables/Demos/FilterDemos.cpp
@@ -1,12 +1,12 @@
 #include "FilterDemos.h"
-#include "../../Modules/RAPT.cpp"
-#include "../Shared/Plotting/Plotting.cpp"
+#include "../Shared/Plotting/Plotting.cpp"  // templates from there will be instantiated here
 
 void ladderImpulseResponse()
 {
   // Demonstrates, how to use the LadderFilter class. 
 
-  typedef float Real; // real numbers are "float" - you can change this to "double"
+  //typedef float Real; // real numbers are "float" - you can change this to "double"
+  typedef double Real; // real numbers are "float" - you can change this to "double"
   typedef RAPT::LadderFilter<Real, Real> LadderFilterR;  // for convenience
 
   // create and set up the filter:
